TYP: move ``matrix`` from ``__init__.pyi`` to ``matrixlib/defmatrix.pyi``
