+++ numpy/__init__.pyi
@@ -606,6 +606,7 @@ from numpy.lib._utils_impl import (
 from numpy.matrixlib import (
     asmatrix,
     bmat,
+    matrix,
 )
 
 __all__ = [  # noqa: RUF022
@@ -5985,192 +5986,6 @@ class poly1d:
         m: SupportsInt | SupportsIndex = 1,
         k: _ArrayLikeComplex_co | _ArrayLikeObject_co | None = 0,
     ) -> poly1d: ...
-
-class matrix(ndarray[_2DShapeT_co, _DTypeT_co]):
-    __array_priority__: ClassVar[float] = 10.0  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    def __new__(
-        subtype,  # pyright: ignore[reportSelfClsParameterName]
-        data: ArrayLike,
-        dtype: DTypeLike | None = None,
-        copy: builtins.bool = True,
-    ) -> matrix[_2D, Incomplete]: ...
-    def __array_finalize__(self, obj: object) -> None: ...
-
-    @overload  # type: ignore[override]
-    def __getitem__(
-        self, key: SupportsIndex | _ArrayLikeInt_co | tuple[SupportsIndex | _ArrayLikeInt_co, ...], /
-    ) -> Incomplete: ...
-    @overload
-    def __getitem__(self, key: _ToIndices, /) -> matrix[_2D, _DTypeT_co]: ...
-    @overload
-    def __getitem__(self: matrix[Any, dtype[void]], key: str, /) -> matrix[_2D, dtype]: ...
-    @overload
-    def __getitem__(self: matrix[Any, dtype[void]], key: list[str], /) -> matrix[_2DShapeT_co, _DTypeT_co]: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    #
-    def __mul__(self, other: ArrayLike, /) -> matrix[_2D, Incomplete]: ...  # type: ignore[override]  # pyright: ignore[reportIncompatibleMethodOverride]
-    def __rmul__(self, other: ArrayLike, /) -> matrix[_2D, Incomplete]: ...  # type: ignore[override]  # pyright: ignore[reportIncompatibleMethodOverride]
-    def __imul__(self, other: ArrayLike, /) -> Self: ...
-
-    #
-    def __pow__(self, other: ArrayLike, /) -> matrix[_2D, Incomplete]: ...  # type: ignore[override]  # pyright: ignore[reportIncompatibleMethodOverride]
-    def __rpow__(self, other: ArrayLike, /) -> matrix[_2D, Incomplete]: ...  # type: ignore[override]  # pyright: ignore[reportIncompatibleMethodOverride]
-    def __ipow__(self, other: ArrayLike, /) -> Self: ...  # type: ignore[override]
-
-    # keep in sync with `prod` and `mean`
-    @overload  # type: ignore[override]
-    def sum(self, axis: None = None, dtype: DTypeLike | None = None, out: None = None) -> Incomplete: ...
-    @overload
-    def sum(self, axis: _ShapeLike, dtype: DTypeLike | None = None, out: None = None) -> matrix[_2D, Incomplete]: ...
-    @overload
-    def sum(self, axis: _ShapeLike | None, dtype: DTypeLike | None, out: _ArrayT) -> _ArrayT: ...
-    @overload
-    def sum(self, axis: _ShapeLike | None = None, dtype: DTypeLike | None = None, *, out: _ArrayT) -> _ArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `sum` and `mean`
-    @overload  # type: ignore[override]
-    def prod(self, axis: None = None, dtype: DTypeLike | None = None, out: None = None) -> Incomplete: ...
-    @overload
-    def prod(self, axis: _ShapeLike, dtype: DTypeLike | None = None, out: None = None) -> matrix[_2D, Incomplete]: ...
-    @overload
-    def prod(self, axis: _ShapeLike | None, dtype: DTypeLike | None, out: _ArrayT) -> _ArrayT: ...
-    @overload
-    def prod(self, axis: _ShapeLike | None = None, dtype: DTypeLike | None = None, *, out: _ArrayT) -> _ArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `sum` and `prod`
-    @overload  # type: ignore[override]
-    def mean(self, axis: None = None, dtype: DTypeLike | None = None, out: None = None) -> Incomplete: ...
-    @overload
-    def mean(self, axis: _ShapeLike, dtype: DTypeLike | None = None, out: None = None) -> matrix[_2D, Incomplete]: ...
-    @overload
-    def mean(self, axis: _ShapeLike | None, dtype: DTypeLike | None, out: _ArrayT) -> _ArrayT: ...
-    @overload
-    def mean(self, axis: _ShapeLike | None = None, dtype: DTypeLike | None = None, *, out: _ArrayT) -> _ArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `var`
-    @overload  # type: ignore[override]
-    def std(self, axis: None = None, dtype: DTypeLike | None = None, out: None = None, ddof: float = 0) -> Incomplete: ...
-    @overload
-    def std(
-        self, axis: _ShapeLike, dtype: DTypeLike | None = None, out: None = None, ddof: float = 0
-    ) -> matrix[_2D, Incomplete]: ...
-    @overload
-    def std(self, axis: _ShapeLike | None, dtype: DTypeLike | None, out: _ArrayT, ddof: float = 0) -> _ArrayT: ...
-    @overload
-    def std(  # pyright: ignore[reportIncompatibleMethodOverride]
-        self, axis: _ShapeLike | None = None, dtype: DTypeLike | None = None, *, out: _ArrayT, ddof: float = 0
-    ) -> _ArrayT: ...
-
-    # keep in sync with `std`
-    @overload  # type: ignore[override]
-    def var(self, axis: None = None, dtype: DTypeLike | None = None, out: None = None, ddof: float = 0) -> Incomplete: ...
-    @overload
-    def var(
-        self, axis: _ShapeLike, dtype: DTypeLike | None = None, out: None = None, ddof: float = 0
-    ) -> matrix[_2D, Incomplete]: ...
-    @overload
-    def var(self, axis: _ShapeLike | None, dtype: DTypeLike | None, out: _ArrayT, ddof: float = 0) -> _ArrayT: ...
-    @overload
-    def var(  # pyright: ignore[reportIncompatibleMethodOverride]
-        self, axis: _ShapeLike | None = None, dtype: DTypeLike | None = None, *, out: _ArrayT, ddof: float = 0
-    ) -> _ArrayT: ...
-
-    # keep in sync with `all`
-    @overload  # type: ignore[override]
-    def any(self, axis: None = None, out: None = None) -> np.bool: ...
-    @overload
-    def any(self, axis: _ShapeLike, out: None = None) -> matrix[_2D, dtype[np.bool]]: ...
-    @overload
-    def any(self, axis: _ShapeLike | None, out: _ArrayT) -> _ArrayT: ...
-    @overload
-    def any(self, axis: _ShapeLike | None = None, *, out: _ArrayT) -> _ArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `any`
-    @overload  # type: ignore[override]
-    def all(self, axis: None = None, out: None = None) -> np.bool: ...
-    @overload
-    def all(self, axis: _ShapeLike, out: None = None) -> matrix[_2D, dtype[np.bool]]: ...
-    @overload
-    def all(self, axis: _ShapeLike | None, out: _ArrayT) -> _ArrayT: ...
-    @overload
-    def all(self, axis: _ShapeLike | None = None, *, out: _ArrayT) -> _ArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `min` and `ptp`
-    @overload  # type: ignore[override]
-    def max(self: NDArray[_ScalarT], axis: None = None, out: None = None) -> _ScalarT: ...
-    @overload
-    def max(self, axis: _ShapeLike, out: None = None) -> matrix[_2D, _DTypeT_co]: ...
-    @overload
-    def max(self, axis: _ShapeLike | None, out: _ArrayT) -> _ArrayT: ...
-    @overload
-    def max(self, axis: _ShapeLike | None = None, *, out: _ArrayT) -> _ArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `max` and `ptp`
-    @overload  # type: ignore[override]
-    def min(self: NDArray[_ScalarT], axis: None = None, out: None = None) -> _ScalarT: ...
-    @overload
-    def min(self, axis: _ShapeLike, out: None = None) -> matrix[_2D, _DTypeT_co]: ...
-    @overload
-    def min(self, axis: _ShapeLike | None, out: _ArrayT) -> _ArrayT: ...
-    @overload
-    def min(self, axis: _ShapeLike | None = None, *, out: _ArrayT) -> _ArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `max` and `min`
-    @overload
-    def ptp(self: NDArray[_ScalarT], axis: None = None, out: None = None) -> _ScalarT: ...
-    @overload
-    def ptp(self, axis: _ShapeLike, out: None = None) -> matrix[_2D, _DTypeT_co]: ...
-    @overload
-    def ptp(self, axis: _ShapeLike | None, out: _ArrayT) -> _ArrayT: ...
-    @overload
-    def ptp(self, axis: _ShapeLike | None = None, *, out: _ArrayT) -> _ArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `argmin`
-    @overload  # type: ignore[override]
-    def argmax(self: NDArray[_ScalarT], axis: None = None, out: None = None) -> intp: ...
-    @overload
-    def argmax(self, axis: _ShapeLike, out: None = None) -> matrix[_2D, dtype[intp]]: ...
-    @overload
-    def argmax(self, axis: _ShapeLike | None, out: _BoolOrIntArrayT) -> _BoolOrIntArrayT: ...
-    @overload
-    def argmax(self, axis: _ShapeLike | None = None, *, out: _BoolOrIntArrayT) -> _BoolOrIntArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # keep in sync with `argmax`
-    @overload  # type: ignore[override]
-    def argmin(self: NDArray[_ScalarT], axis: None = None, out: None = None) -> intp: ...
-    @overload
-    def argmin(self, axis: _ShapeLike, out: None = None) -> matrix[_2D, dtype[intp]]: ...
-    @overload
-    def argmin(self, axis: _ShapeLike | None, out: _BoolOrIntArrayT) -> _BoolOrIntArrayT: ...
-    @overload
-    def argmin(self, axis: _ShapeLike | None = None, *, out: _BoolOrIntArrayT) -> _BoolOrIntArrayT: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    #the second overload handles the (rare) case that the matrix is not 2-d
-    @overload
-    def tolist(self: matrix[_2D, dtype[generic[_T]]]) -> list[list[_T]]: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-    @overload
-    def tolist(self) -> Incomplete: ...  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # these three methods will at least return a `2-d` array of shape (1, n)
-    def squeeze(self, axis: _ShapeLike | None = None) -> matrix[_2D, _DTypeT_co]: ...
-    def ravel(self, /, order: _OrderKACF = "C") -> matrix[_2D, _DTypeT_co]: ...  # type: ignore[override]  # pyright: ignore[reportIncompatibleMethodOverride]
-    def flatten(self, /, order: _OrderKACF = "C") -> matrix[_2D, _DTypeT_co]: ...  # type: ignore[override]  # pyright: ignore[reportIncompatibleMethodOverride]
-
-    # matrix.T is inherited from _ScalarOrArrayCommon
-    def getT(self) -> Self: ...
-    @property
-    def I(self) -> matrix[_2D, Incomplete]: ...  # noqa: E743
-    def getI(self) -> matrix[_2D, Incomplete]: ...
-    @property
-    def A(self) -> ndarray[_2DShapeT_co, _DTypeT_co]: ...
-    def getA(self) -> ndarray[_2DShapeT_co, _DTypeT_co]: ...
-    @property
-    def A1(self) -> ndarray[_AnyShape, _DTypeT_co]: ...
-    def getA1(self) -> ndarray[_AnyShape, _DTypeT_co]: ...
-    @property
-    def H(self) -> matrix[_2D, _DTypeT_co]: ...
-    def getH(self) -> matrix[_2D, _DTypeT_co]: ...
 
 def from_dlpack(
     x: _SupportsDLPack[None],
